requirements.txt
bs4
pymediainfo
gtts
pydub
cinemagoer
edge-tts
emoji
geopy
git-python
glitch_this
googlesearch-python
googletrans @ git+https://github.com/badmunda98/py-googletrans
heroku3
httpx
lottie
lyricsgenius
motor
pillow
psutil
pyroaddon
pyrogram
pySmartDL
python-dotenv
pytz
requests
selenium
speedtest-cli
telegraph
tgcrypto
Unidecode
wikipedia
py-tgcalls
yt-dlp @ git+https://github.com/Badmunda98/yt-dp-fork
flask_restful
flask
gunicorn